requirements.txt
gigachain==0.1.16
gigachain_community==0.0.34
gigachain_core==0.1.45
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.33.0
streamlit_extras==0.4.2
tiktoken==0.6.0
